Create classifieds page
This commit changes the original name of "Ads" to "Classifieds". To
achieve this, the webmaster will have to create yml files for each ad
with simple front matter that can include the NSIP ID numbers of any
animals for sale. Breeders can submit the information using a google
form. The code matches the breeder's name or that of the farm to the
breeders yml files. The output indicates when the ad was placed and has
the ads expire after (currently) three months. Because this is a static
site, the repository needs an empty commit to refresh the GitHub pages
version. Closes #4.